+++ src/writer/counting_writer.rs
