+++ src/writer/seq_reader.rs
@@ -1,8 +1,4 @@
-#[cfg(feature = "util")]
-use std::path::PathBuf;
-use std::{fs::File, io::Read, ops::Deref};
-
-use crc32fast::Hasher;
+use std::{io::Read, ops::Deref};
 
 pub(crate) struct SeqReader<R> {
     readers: Vec<R>,
@@ -44,89 +40,5 @@ impl<R: Read> Read for SeqReader<R> {
         }
 
         Ok(i)
-    }
-}
-
-pub struct SourceReader<R> {
-    reader: R,
-    size: usize,
-    crc: Hasher,
-    crc_value: u32,
-}
-
-impl<R> From<R> for SourceReader<R> {
-    fn from(value: R) -> Self {
-        Self::new(value)
-    }
-}
-
-impl<R: Read> Read for SourceReader<R> {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        let n = self.reader.read(buf)?;
-        if self.crc_value == 0 {
-            if n > 0 {
-                self.size += n;
-                self.crc.update(&buf[..n]);
-            } else {
-                let crc = std::mem::replace(&mut self.crc, Hasher::new());
-                self.crc_value = crc.finalize();
-            }
-        }
-        Ok(n)
-    }
-}
-
-impl<R> SourceReader<R> {
-    pub fn new(reader: R) -> Self {
-        Self {
-            reader,
-            size: 0,
-            crc: Hasher::new(),
-            crc_value: 0,
-        }
-    }
-
-    pub fn read_count(&self) -> usize {
-        self.size
-    }
-
-    pub fn crc_value(&self) -> u32 {
-        self.crc_value
-    }
-}
-
-#[cfg(feature = "util")]
-pub(crate) struct LazyFileReader {
-    path: PathBuf,
-    reader: Option<File>,
-    end: bool,
-}
-
-#[cfg(feature = "util")]
-impl LazyFileReader {
-    pub fn new(path: PathBuf) -> Self {
-        Self {
-            path,
-            reader: None,
-            end: false,
-        }
-    }
-}
-
-#[cfg(feature = "util")]
-impl Read for LazyFileReader {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        if self.end {
-            return Ok(0);
-        }
-        if self.reader.is_none() {
-            self.reader = Some(File::open(&self.path)?);
-        }
-        let n = self.reader.as_mut().unwrap().read(buf)?;
-        if n == 0 {
-            self.end = true;
-            self.reader = None;
-        }
-        Ok(n)
     }
 }
