+++ tests/compress_tests.rs
@@ -4,8 +4,7 @@ use std::{
     io::{Cursor, Read},
 };
 
-use sevenz_rust2::method_options::*;
-use sevenz_rust2::*;
+use sevenz_rust2::{method_options::*, *};
 use tempfile::*;
 
 #[cfg(feature = "compress")]
